[hyrrokkin] Don't fail if publicsuffix.org is down.
The suffix list is redownloaded from publicsuffix.org every week to keep
it up-to-date, but if that download failed in any way, an empty list was
written to disk and then used subsequently, resulting in failures to
identify any domains.

Now, there is a version of the list supplied with the gem. That version
is used initially and will be updated on a weekly basis thereafter.